use binary package name for Debian installation
For the Debian distribution, the README indicates `browserpass` as the name of the package to be used by `apt`. It turns out that `browserpass` is the name of the _source package_. But the _binary package_, the one which can be installed by `apt` and `dpkg` is called, `webext-browserpass`.

This proposed modification reflects this.